pyproject.toml
[tool.poetry]
name = "starpilot"
version = "0.1.0"
description = "Discover the hidden power of your GitHub Stars"
authors = ["Dave Parr <8363743+DaveParr@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.scripts]
starpilot = "starpilot.main:app"

[tool.poetry.dependencies]
python = "^3.10"
typer = {extras = ["all"], version = "^0.9.0"}
langchain = "^0.0.331"
gpt4all = "^2.0.2"
chromadb = "^0.4.15"
pygithub = "^2.1.1"
unstructured = "^0.10.29"
markdown = "^3.5.1"

[tool.poetry.group.dev.dependencies]
black = "^23.10.1"
isort = "^5.12.0"
ruff = "^0.1.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
